fix: lint

+++ api/v1/server/oas/transformers/v1/workflow_runs.go
@@ -107,7 +107,7 @@ func ToWorkflowRunMany(
 		actionId := ""
 
 		if task.TaskId != nil {
-			actionIdFromMap, ok := taskIdToActionId[int64(*task.TaskId)]
+			actionIdFromMap, ok := taskIdToActionId[*task.TaskId]
 
 			if ok {
 				actionId = actionIdFromMap
